doc + settings fix
